+++ discovery/service/kafka_rest.py
@@ -211,7 +211,7 @@ class KafkaRestServicePropertyBaseBuilder(AbstractPropertyBuilder):
     def _build_log4j_properties(self, service_properties: dict) -> tuple:
         log4j_file = self.get_log_file_path(self.input_context, self.service, self.hosts, "KAFKAREST_LOG4J_OPTS")
         default_log4j_file = "/etc/kafka-rest/log4j.properties"
-        root_logger, file = self.get_root_logger(self.input_context, self.service, self.hosts, log4j_file, default_log4j_file)
+        root_logger, file = self.get_root_logger(self.input_context, self.hosts, log4j_file, default_log4j_file)
 
         if root_logger is None or file is None:
             return self.group, {'kafka_rest_custom_log4j': False}
