+++ discovery/service/service.py
@@ -1,14 +1,11 @@
 import abc
-import base64
 import re
 from abc import ABC
-import string
-import json
 
-from discovery.system.system import SystemPropertyManager
-from discovery.utils.constants import ConfluentServices, DEFAULT_KEY
+from discovery.api.manager import ServicePropertyManager
+from discovery.utils.constants import ConfluentServices
 from discovery.utils.inventory import CPInventoryManager
-from discovery.utils.utils import InputContext, PythonAPIUtils, load_properties_to_dict, Logger
+from discovery.utils.utils import InputContext, Logger
 
 logger = Logger.get_logger()
 
@@ -19,7 +16,24 @@ class AbstractPropertyBuilder(ABC):
     def build_properties(self):
         pass
 
-    def get_service_host(self, service: ConfluentServices, inventory: CPInventoryManager):
+    @staticmethod
+    def get_keystore_alias_names(input_context: InputContext, hosts: list, keystorepass: str, keystorepath: str):
+        return ServicePropertyManager.get_keystore_alias_names(input_context, hosts, keystorepass, keystorepath)
+
+    @staticmethod
+    def get_root_logger(input_context: InputContext, hosts: list, log4j_file, default_log4j_file):
+        return ServicePropertyManager.get_root_logger(input_context, hosts, log4j_file, default_log4j_file)
+
+    @staticmethod
+    def get_log_file_path(input_context: InputContext, service: ConfluentServices, hosts: list, log4j_opts_env_var):
+        return ServicePropertyManager.get_log_file_path(input_context, service, hosts, log4j_opts_env_var)
+
+    @staticmethod
+    def get_property_mappings(input_context: InputContext, service: ConfluentServices, hosts: list):
+        return ServicePropertyManager.get_property_mappings(input_context, service, hosts)
+
+    @staticmethod
+    def get_service_host(service: ConfluentServices, inventory: CPInventoryManager):
         group_name = service.value.get("group")
         hosts = inventory.get_groups_dict().get(group_name)
 
@@ -34,20 +48,20 @@ class AbstractPropertyBuilder(ABC):
 
     @staticmethod
     def get_secret_protection_key(input_context: InputContext, service: ConfluentServices, hosts: list):
-        env_details = AbstractPropertyBuilder._get_env_details(input_context, service, hosts)
+        env_details = ServicePropertyManager.get_env_details(input_context, service, hosts)
         return env_details.get("CONFLUENT_SECURITY_MASTER_KEY", None)
 
     @staticmethod
     def get_rocksdb_path(input_context: InputContext, service: ConfluentServices, hosts: list):
-        env_details = AbstractPropertyBuilder._get_env_details(input_context, service, hosts)
+        env_details = ServicePropertyManager.get_env_details(input_context, service, hosts)
         return env_details.get("ROCKSDB_SHAREDLIB_DIR", "")
 
     @staticmethod
     def get_jvm_arguments(input_context: InputContext, service: ConfluentServices, hosts: list):
         # Build Java runtime overrides
-        env_details = AbstractPropertyBuilder._get_env_details(input_context, service, hosts)
+        env_details = ServicePropertyManager.get_env_details(input_context, service, hosts)
         heap_ops = env_details.get('KAFKA_HEAP_OPTS', '')
-        kafka_ops= env_details.get('KAFKA_OPTS', '')
+        kafka_ops = env_details.get('KAFKA_OPTS', '')
         # Remove java agent configurations. These will be populated by other configs.
         kafka_ops = re.sub('-javaagent.*?( |$)', '', kafka_ops).strip()
 
@@ -59,203 +73,6 @@ class AbstractPropertyBuilder(ABC):
             jvm_str = f"{jvm_str} {kafka_ops}"
 
         return jvm_str
-
-    @staticmethod
-    def _get_env_details(input_context: InputContext, service: ConfluentServices, hosts: list):
-        service_facts = AbstractPropertyBuilder.get_service_facts(input_context, service, hosts)
-        environment = service_facts.get("Environment", None)
-        return AbstractPropertyBuilder.parse_environment_details(environment)
-
-    @staticmethod
-    def __get_service_properties_file(input_context: InputContext,
-                                      service: ConfluentServices,
-                                      hosts: list):
-        if not hosts:
-            logger.error(f"Host list is empty for service {service.value.get('name')}")
-            return None
-
-        host = hosts[0]
-        service_details = SystemPropertyManager.get_service_details(input_context, service, host)
-        execution_command = service_details.get(host).get("status").get("ExecStart")
-
-        # check if we have flag based configs
-        property_files = dict()
-        matches = re.findall('(--[\w\.]+\.config)*\s+([\w\/-]+\.properties)', execution_command)
-        for match in matches:
-            key, path = match
-            key = key.strip('--') if key else DEFAULT_KEY
-            property_files[key] = path
-
-        if not property_files:
-            logger.error(f"Cannot find associated properties file for service {service.value.get('name')}")
-        return property_files
-
-    @staticmethod
-    def get_property_mappings(input_context: InputContext, service: ConfluentServices, hosts: list) -> dict:
-
-        mappings = dict()
-        property_file_dict = AbstractPropertyBuilder.__get_service_properties_file(input_context, service, hosts)
-        if not property_file_dict:
-            logger.error(f"Could not get the service seed property file.")
-            return mappings
-
-        for key, file in property_file_dict.items():
-            play = dict(
-                name="Slurp properties file",
-                hosts=hosts,
-                gather_facts='no',
-                tasks=[
-                    dict(action=dict(module='slurp', args=dict(src=file)))
-                ]
-            )
-            response = PythonAPIUtils.execute_play(input_context, play)
-            for host in hosts:
-                properties = base64.b64decode(response[host]._result['content']).decode('utf-8')
-                host_properites = mappings.get(host, dict())
-                host_properites.update({key: load_properties_to_dict(properties)})
-                mappings[host] = host_properites
-
-        return mappings
-
-    @staticmethod
-    def get_keystore_alias_names(input_context: InputContext, hosts: list, keystorepass: str, keystorepath: str) -> str:
-
-        if not keystorepath or not keystorepass:
-            return []
-
-        play = dict(
-            name="Get keystore alias name",
-            hosts=hosts,
-            gather_facts='no',
-            tasks=[
-                dict(action=dict(module='shell',
-                                 args=f"keytool -list -v -storepass {keystorepass} -keystore "
-                                      f"{keystorepath} | grep 'Alias name' "
-                                      "| awk '{print $3}'"))
-            ]
-        )
-
-        response = PythonAPIUtils.execute_play(input_context, play)
-        if response[hosts[0]]._result['rc'] == 0:
-            stdout = response[hosts[0]]._result['stdout']
-            aliases = [y for y in (x.strip() for x in stdout.splitlines()) if y]
-            return aliases
-        else:
-            return []
-
-    @staticmethod
-    def get_log_file_path(input_context: InputContext, service: ConfluentServices, hosts: list, log4j_opts_env_var):
-        # check if overriden as env var
-        env_details = AbstractPropertyBuilder._get_env_details(input_context, service, hosts)
-        log4j_opts = env_details.get(log4j_opts_env_var, None)
-        if log4j_opts is not None:
-            if "-Dlog4j.configuration=file:" in log4j_opts:
-                log4j_path = log4j_opts.split("-Dlog4j.configuration=file:",1)[1].split(" ",1)[0]
-                return log4j_path
-
-        # if not overridden, read from java process details
-        play = dict(
-            name="Get java process details for cp component",
-            hosts=hosts,
-            gather_facts='no',
-            tasks=[
-                dict(action=dict(module='shell',
-                                 args=f"ps aux | grep java | grep log4j"))
-            ]
-        )
-        response = PythonAPIUtils.execute_play(input_context, play)
-        if len(response)==0:
-            return None
-        if response[hosts[0]]._result['rc'] == 0:
-            process_details = response[hosts[0]]._result['stdout']
-            if "-Dlog4j.configuration=file:" in process_details:
-                log4j_path = process_details.split("-Dlog4j.configuration=file:",1)[1].split(" ",1)[0]
-                return log4j_path
-
-        return None
-
-    @staticmethod
-    def get_root_logger(input_context: InputContext, service: ConfluentServices, hosts: list, log4j_file, default_log4j_file):
-        for file in [log4j_file, default_log4j_file]:
-            if file is None:
-                continue
-            play = dict(
-                name="Get root logger definition from log4j file",
-                hosts=hosts,
-                gather_facts='no',
-                tasks=[
-                    dict(action=dict(module='shell',
-                                    args=f"grep ^log4j.rootLogger {file}"))
-                ]
-            )
-            response = PythonAPIUtils.execute_play(input_context, play)
-            if len(response)==0:
-                continue
-            if response[hosts[0]]._result['rc'] == 0:
-                logger_def = response[hosts[0]]._result['stdout']
-                return logger_def.split("log4j.rootLogger=",1)[1].strip(), file
-
-        return None, None
-
-    @staticmethod
-    def get_audit_log_properties(input_context: InputContext, hosts: string, mds_user: str, mds_password: str) -> str:
-        # returns list of all clusters in the registry
-        play1 = dict(
-            name="Get list of clusters",
-            hosts=hosts,
-            gather_facts='no',
-            tasks=[
-                dict(action=dict(module='shell',
-                                 args=f"curl -ku '{mds_user}:{mds_password}' "
-                                      f"https://localhost:8090/security/1.0/registry/clusters"))
-            ]
-        )
-
-        response = PythonAPIUtils.execute_play(input_context, play1)
-        if len(response)==0:
-            return "", ""
-        cluster_list = ""
-        if response[hosts]._result['rc'] == 0:
-            cluster_list = response[hosts]._result['stdout']
-        if cluster_list == "":
-            return "", ""
-        try:
-            res = json.loads(cluster_list)
-        except:
-            res = ""
-
-        for cluster in res:
-            cluster_name = cluster['clusterName']
-            payload = f'{{"clusterName":"{cluster_name}"}}'
-            # Look up the Principals who have the given role ResourceOwner on the specified resource (Topic:confluent-audit-log-events) for the given cluster.
-            play2 = dict(
-                name="Check rolebinding to get principal name",
-                hosts=hosts,
-                gather_facts='no',
-                tasks=[
-                    dict(action=dict(module='shell',
-                                    args=f"curl -ku '{mds_user}:{mds_password}' "
-                                        f"https://localhost:8090/security/1.0/lookup/role/ResourceOwner/resource/Topic/name/confluent-audit-log-events "
-                                        f"-H 'Content-Type: application/json' "
-                                        f"-d '{payload}'"))
-                ]
-            )
-            response2 = PythonAPIUtils.execute_play(input_context, play2)
-            if len(response2)==0:
-                continue
-            if response2[hosts]._result['rc'] == 0:
-                stdout = response2[hosts]._result['stdout']
-                try:
-                    role = json.loads(stdout)
-                except:
-                    role = []
-                if len(role) == 0:
-                    continue
-                principal_name = role[0].split(";")[0].split("User:")[1]
-                # If a principal exists for this cluster, this is the destination kafka cluster for audit logs
-                return cluster, principal_name
-
-        return "", ""
 
     @staticmethod
     def build_telemetry_properties(service_prop: dict) -> dict:
@@ -270,17 +87,6 @@ class AbstractPropertyBuilder(ABC):
                 property_dict['telemetry_proxy_username'] = service_prop.get('confluent.telemetry.proxy.username')
                 property_dict['telemetry_proxy_password'] = service_prop.get('confluent.telemetry.proxy.password')
         return property_dict
-
-    @staticmethod
-    def parse_environment_details(env_command: str) -> dict:
-        env_details = dict()
-        tokens = ['KAFKA_HEAP_OPTS', 'KAFKA_OPTS', 'KAFKA_LOG4J_OPTS', 'LOG_DIR']
-        for token in tokens:
-            pattern = f"{token}=(.*?) ([A-Z]{{3}}|$)"
-            match = re.search(pattern, env_command)
-            if match:
-                env_details[token] = match.group(1).rstrip()
-        return env_details
 
     @staticmethod
     def get_service_facts(input_context: InputContext, service: ConfluentServices, hosts: list) -> dict:
@@ -302,12 +108,12 @@ class AbstractPropertyBuilder(ABC):
         user = service_facts.get("User", None)
         group = service_facts.get("Group", None)
         environment = service_facts.get("Environment", None)
-        env_details = AbstractPropertyBuilder.parse_environment_details(environment)
+        env_details = ServicePropertyManager.parse_environment_details(environment)
 
         # Useful information for future usages
-        service_file = service_facts.get("FragmentPath", None)
-        service_override = service_facts.get("DropInPaths", None)
-        description = service_facts.get("Description", None)
+        # service_file = service_facts.get("FragmentPath", None)
+        # service_override = service_facts.get("DropInPaths", None)
+        # description = service_facts.get("Description", None)
 
         service_group = service.value.get("group")
         return service_group, {
@@ -376,12 +182,11 @@ class AbstractPropertyBuilder(ABC):
                 user_dict[key] = value
         return user_dict
 
-
     @staticmethod
-    def get_monitoring_details(input_context, service, hosts, key)->dict:
+    def get_monitoring_details(input_context, service, hosts, key) -> dict:
         monitoring_props = dict()
-        env_details = AbstractPropertyBuilder._get_env_details(input_context, service, hosts)
-        ops_str= env_details.get(key, '')
+        env_details = ServicePropertyManager.get_env_details(input_context, service, hosts)
+        ops_str = env_details.get(key, '')
 
         if 'jolokia.jar' in ops_str:
             monitoring_props['jolokia_enabled'] = True
@@ -395,6 +200,7 @@ class AbstractPropertyBuilder(ABC):
                 monitoring_props['jmxexporter_port'] = int(match.group(1))
 
         return monitoring_props
+
 
 class ServicePropertyBuilder:
     inventory = None
